// src/frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export interface PaintingInfo {
    title: string;
    description: string;
    image: ExternalBlob;
    priceUSD: number;
    dimensions: string;
    material: string;
}
export interface Painting {
    id: bigint;
    title: string;
    description: string;
    image: ExternalBlob;
    priceUSD: number;
    dimensions: string;
    material: string;
}
export interface backendInterface {
    addPainting(info: PaintingInfo): Promise<bigint>;
    deletePainting(id: bigint): Promise<void>;
    getAllPaintings(): Promise<Array<Painting>>;
    getPainting(id: bigint): Promise<Painting>;
    updatePainting(id: bigint, newInfo: PaintingInfo): Promise<void>;
}